byokg-rag/src/graphrag_toolkit/byokg_rag/requirements.txt
pydantic>=2.8.2
boto3
xmltodict
colorama
pyyaml
pytest
faiss-cpu==1.9.0
thefuzz
langchain_huggingface
ipykernel
torch
